limit polynomial correction
piolynomial correction is determined for an interval in hue angle
It is important to limit the polynomial correction outside this
interval to avoid spurious values.
This is done by setting a fixed correction outisde the interval.
At the moment the interval is >37 deg. for all sensors as taken from
Woerd and Wernand 2015 and 2018 papers.